+++ authentication/src/infrastructure/database/postgres/migrations/20200901142257_create_user_view.ts
@@ -5,7 +5,10 @@ const TABLE_NAME = 'user_view';
 
 export async function up(knex: Knex): Promise<void> {
   await knex.schema.createTable(TABLE_NAME, (table: TableBuilder) => {
-    table.uuid('id').primary().notNullable().comment('Unique user id');
+    table.uuid('userId').primary().notNullable();
+    table.string('email', 255).notNullable();
+    table.timestamp('createdAt');
+    table.uuid('sessionId');
   });
 }