+++ common/src/api-contracts/authentication/authentication-query-model.ts
@@ -1,0 +1,6 @@
+export interface UserView {
+  userId: string;
+  email: string;
+  createdAt: string;
+  sessionId: string | undefined;
+}